Centralize Storyblok cache version handling
